fix os; move xccode
