+++ .github/workflows/run_ci.sh
@@ -16,6 +16,7 @@ pip install -U jax
 
 pip install trexio
 pip install mcfun
+pip install --no-deps pyscf-dispersion==1.5.0
 
 # TODO: check if pyscf code is changed using dist-info file
 #pip uninstall -y pyscf-forge
