--- pyscf/dh/util/__init__.py
@@ -1,6 +1,0 @@
-"""
-Utilities for pyscf.dh
-"""
-
-from pyscf.dh.util.xccode.xccode import XCInfo, XCList, XCDH
-from pyscf.dh.util.xccode.xctype import XCType
