+++ pyscf/dh/xccode/__init__.py
@@ -1,0 +1,3 @@
+from pyscf.dh.xccode.xccode import XCInfo, XCList, XCDH
+from pyscf.dh.xccode.xctype import XCType
+from pyscf.dh.xccode.parse import parse_xc_dh, xc_equal
